requirements.txt
lxml>=4.9.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
